Remove error notifs from FileSystem, Separate notification system, added notification bar to greet window.
